wireless-regdb: Permit lower 6 GHz band for Kazakhstan (KZ)
By Order of the Minister of Digital Development No. 564/NK dated
September 19, 2024 [1], an amendment was made to the assignment of
frequency bands, opening up the lower part of the 6 GHz band (5,945 MHz
to 6,425 MHz) was opened up to "SRD (wideband data transmission systems)
WAS/RLAN Data transmission systems/radio local area networks".

The power is limited to 200 mW e.i.r.p. with a spectral density of 10
mW/MHz. Only indoor use is permitted.

While at it, also change the power limits to use mW units to match the
official documents. The 5250-5350 MHz band was left alone, as the
original submission [2] suggests the limit was halved to comply with
TPC requirements.

Initial information was provided by the Wi-Fi Alliance website, which
has a map [3] showing all countries that have enabled 6 GHz WiFi, and
links to official documents. This document was not sufficient as it did
not include power limits or other restrictions.

A subsequent search found a blog post [4] providing details and titles
of the official documents in English. An AI agent was used to search
for the original text of said order [1] after rummaging through
government websites in Russian and Kazakh failed to find anything.

[1] https://adilet.zan.kz/rus/docs/V2400035100
[2] https://lore.kernel.org/wireless-regdb/20191029144815.GR30813@ubuntu-xps13/
[3] https://www.wi-fi.org/regulations-enabling-6-ghz-wi-fi
[4] https://www.onewindow.pro/blog-posts/6-ghz-band-regulation-in-the-republic-of-kazakhstan

Reviewed-by: Ping-Ke Shih <pkshih@realtek.com>
Link: https://lore.kernel.org/r/20250930042948.427154-1-wens@kernel.org
Signed-off-by: Chen-Yu Tsai <wens@kernel.org>